requirements.txt
python-chess>=1.999
pytest>=8.0
requests>=2.31.0
fastapi>=0.115.0
uvicorn>=0.30.0
